+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Made the pattern argument optional by adding `nargs="?"`
+
 ## [0.3.9] - 2025-10-25
 
 ### Added